+++ GeoQuizChallenge/app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/AppTheme"
         >
 
-        <activity android:name=".Views.QuizActivity">
+        <activity android:name=".views.QuizActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
