+++ GeoQuizChallenge/app/src/main/res/layout-land/activity_quiz.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".Views.QuizActivity"
+    tools:context=".views.QuizActivity"
     >
 
     <TextView
